userspace: split policy lowering into policies_lower.go; residual core
Move the policy-lowering concern verbatim from policies.go into a new
policies_lower.go: buildPolicySnapshots,
buildPolicySnapshotsWithSchedulerState,
buildPolicySnapshotsWithSchedulerStateAndFeeds, buildOneRuleSnapshot, and
policyActionString. This is the Junos-policy-config -> PolicyRuleSnapshot
lowering path (#2124/#2508/#3148/#3261/#3376).

policies.go is now the shared core the concern files build on: the two
content-rejection sentinels, the policyRuleSlot type + policyID method and
walkPolicyRuleSlots SSOT namespace walk (shared by lowering and the ID
helpers), the isV4CIDR/isV6CIDR predicates (shared by the address-book and
representability concerns), and the hex import-silencer.

This completes the #4421 policies.go split (codex-173 #4): one 1432-line
grab-bag becomes a shared core plus six by-concern siblings (lower, ids,
addrbook, representable, reject, scheduler). Pure code-motion, no logic
change; every declaration is a byte-identical slice of the origin/master
policies.go (verified: identical sorted decl-name set and identical
decl-body multiset across all seven files; go build ./... and
go test ./pkg/dataplane/userspace/... green).
